pyproject.toml
[project]
name = "fenrir-sabot"
version = "0.1.0"
description = "Fenrir Telegram Bot"
authors = [
    {name = "Antares Husky",email = "husky@antares.dog"}
]
license = {text = "gpl-3.0"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.0,<23.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

package-mode = false

[tool.poetry]
packages = [
    {include = "src"}
]



[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
matplotlib = "^3.10.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
